+++ R/print.rd_pat.R
@@ -27,7 +27,7 @@
 #'
 #' @aliases print.rd_pat
 #' @rdname print.rd_pat
-
+#'
 #' @export
 print.rd_pat <- function(x,
                          digits = 4,