+++ R/print.rd_ras.R
@@ -25,7 +25,7 @@
 #'
 #' @aliases print.rd_ras
 #' @rdname print.rd_ras
-
+#'
 #' @export
 print.rd_ras <- function(x,
                          ...) {