+++ R/shar-package.R
@@ -9,6 +9,7 @@
 #'
 #' @name shar
 #' @docType package
+#' @keywords internal
 "_PACKAGE"
 
 # Global variables